feat: Implement logging functionality with enhanced formatting and state management
